+++ server/templates/measures.html
@@ -13,23 +13,24 @@
         {{ region_dropdown("regionDropdown") }}
         <div class="active-infections-block">
           <span class="number-subheader" id="infections-date"></span>
-          <span class="active-infections"
-            >Active Infections:
-            <span class="infections-estimated" id="infections-estimated"
-              >&mdash;</span
-            >
+          <span class="active-infections">Active Infections:
+            <span class="infections-estimated" id="infections-estimated">
+              &mdash;
+            </span>
           </span>
           <span class="infections-confirmed">
             Confirmed Infections:
-            <span id="infections-confirmed">&mdash;</span></span
-          >
+            <span id="infections-confirmed">
+              &mdash;
+            </span>
+          </span>
         </div>
 
         <div class="population-block">
           <span class="number-subheader">Population</span>
-          <span class="infections-population" id="infections-population"
-            >&mdash;</span
-          >
+          <span class="infections-population" id="infections-population">
+            &mdash;
+          </span>
         </div>
       </div>
